requirements.txt
tinygrad
numpy
hypothesis
ipykernel
chalk-diagrams
